% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{as_df}
\alias{as_df}
\title{Wrapper for as.data.frame}
\usage{
as_df(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{a object you want to make a data frame}

\item{row.names}{NULL or a character vector giving the row names for the data frame. Missing values are not allowed}

\item{optional}{ogical. If TRUE, setting row names and converting column names (to syntactic names: see make.names) is optional. Note that all of R's base package as.data.frame() methods use optional only for column names treatment, basically with the meaning of data.frame(*, check.names = !optional). See also the make.names argument of the matrix method.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
a data frame
}
\description{
Wrapper for as.data.frame
}
\author{
Abram Fleishman \email{abram@conservationmetics.com}
}
