% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{goff}
\alias{goff}
\title{Wrapper for dev.off() and graphics.off()}
\usage{
goff()
}
\description{
Wrapper for dev.off() and graphics.off()
}
\author{
Abram Fleishman \email{abram@conservationmetics.com}
}
