% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens_drones.R
\docType{data}
\name{lens_drones}
\alias{lens_drones}
\title{Access drones patent data from the Lens}
\format{{ A data frame with 9071 observations of 24 variables:
 \describe{
   \item{\code{number}}{The record number (row no), a character vector}
   \item{\code{jurisdiction}}{The two letter country or instrument code where a document was published, a character vector}
   \item{\code{kind}}{The kind code describing the publication level of a document, a character vector}
   \item{\code{publication_number}}{The publication number including forward /, a character vector}
   \item{\code{lens_id}}{The internal Lens database identifier, a character vector}
   \item{\code{publication_date}}{a character vector}
   \item{\code{publication_year}}{a character vector}
   \item{\code{application_number}}{a character vector}
   \item{\code{application_date}}{a character vector}
   \item{\code{priority_numbers}}{Patent priority numbers in long format with dates, a character vector}
   \item{\code{title}}{The patent document title, a character vector}
   \item{\code{applicants}}{patent applicants concatenated with ";;", a character vector}
   \item{\code{inventors}}{The original inventor name, concatenated with ";;", a character vector}
   \item{\code{url}}{url link to the record, a character vector}
   \item{\code{type}}{A Patent Application, Granted Patent or other type, a character vector}
   \item{\code{full_text}}{Indicates whether the full text is available at the Lens, a character vector}
   \item{\code{cited_count}}{Number of later patent filings citing the document, a character vector}
   \item{\code{simple family size}}{Assumed to be the DOCDB family based on shared priority, a character vector}
   \item{\code{extented_family_size}}{Assumed to be the extended INPADOC family but not documented by the Lens, a character vector}
   \item{\code{sequence_count}}{A count of DNA or amino acid sequences, a character vector}
   \item{\code{cpc_classification}}{The Cooperative Patent Classification Codes, a character vector}
   \item{\code{ipcr_classifications}}{International Patent Classification - IPC - codes, a character vector}
   \item{\code{npl_citation_ids}}{Non patent literature citation identifiers, a character vector}

 }
}}
\source{
\href{https://www.lens.org/lens/collection/45155}{The Lens Patent Databse}
}
\usage{
data("lens_drones")
}
\description{
A set of 9071 patent documents from the open access Lens patent database containing the word drone or drones in the full text of documents published between 1948 and the 31st of December 2017 and reduced from 16,419 publications to one record per family. The records can be accessed online at the public collection \href{https://www.lens.org/lens/collection/45155}{https://www.lens.org/lens/collection/45155}.
}
\details{
Field names in this dataset have been simplified from their original long form to make them easier to work with in R. Lens Patent data fields are concatenated with a double semicolon and require tidying for accurate counts. Applicant and inventor names have not been cleaned.
}
\keyword{datasets}
