% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_one_tile.R
\name{make_one_tile}
\alias{make_one_tile}
\title{Tile an image}
\usage{
make_one_tile(img, crop_path, width, buff, height, corner_x, corner_y)
}
\arguments{
\item{img}{Path to an image folder}

\item{crop_path}{Path to an  folder to be made that will have the images out}

\item{width}{width in pixels of the crop}

\item{buff}{buffer to overlap crop by x pixels}

\item{height}{height in pixels of the crop}

\item{corner_x}{x coord of the top left corner of the crop}

\item{corner_y}{y coord of the top left corner of the crop}
}
\value{
not much really.  just a list of names of images in
}
\description{
Tile images in a directory
}
\author{
Abram Fleishman \email{abram@conservationmetics.com}
}
