% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{na_table}
\alias{na_table}
\title{Wrapper for is.na + table}
\usage{
na_table(x)
}
\arguments{
\item{x}{column you want to count NAs in}
}
\description{
Wrapper for is.na + table
}
\author{
Abram Fleishman \email{abram@conservationmetics.com}
}
