% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_list_from_rows.R
\name{polygon_list_from_rows}
\alias{polygon_list_from_rows}
\title{create list of polygons from a data.frame row}
\usage{
polygon_list_from_rows(dat, xmin = "image_xmin", xmax = "image_xmax",
  ymin = "image_ymin", ymax = "image_ymax")
}
\arguments{
\item{dat}{data.frame with column names specified in x* and y* below}

\item{xmin}{quoted column name}

\item{xmax}{quoted column name}

\item{ymin}{quoted column name}

\item{ymax}{quoted column name}
}
\value{
a list with each entry a sf Polygon object
}
\description{
Take a row from a data.frame and using xmin.xmax,ymin,ymax coordinents make
a list of polygons
}
\author{
Abram Fleishman \email{abram@conservationmetics.com}
}
