% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exif_ssh.R
\name{read_exif_ssh}
\alias{read_exif_ssh}
\title{Read file exif data over ssh}
\usage{
read_exif_ssh(ssh_host = "user@10.74.76.105:22", passwrd,
  nas_image_path = "//NAS1/NAS3_2Mar15/Images/CMI_Office_2015",
  dbox_project_path = "D:/CM,Inc/Dropbox (CMI)/CMI_Team/Analysis/2019/CMI_Office_2015",
  tags = "")
}
\arguments{
\item{ssh_host}{server to ssh into default is 'user at 10.74.76.105:22'}

\item{passwrd}{password to server}

\item{nas_image_path}{path of nas image folder with data to scan}

\item{dbox_project_path}{dropbox project folder to save csv out}

\item{tags}{any additional parameters to add to exiftool call (can be specific tags to get back)}
}
\value{
a data.frame/tibble with all the exif data for the folder you scaned
}
\description{
Reads file exif data from a folder on the nas and saves it as a csv in dropbox.
}
\author{
Abram Fleishman \email{abram@conservationmetics.com}
}
