% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_photos.R
\name{tile_photos}
\alias{tile_photos}
\title{Tile an image}
\usage{

  tile_photos(in_folder = "//NAS1/NAS3_2Mar15/Images/FC_BBAL_Drone_Images/",
  out_folder = "//NAS1/NAS3_2Mar15/Images/FC_BBAL_Drone_Images_tiled",
  width = 1000, height = 600, buff = 30, file_type = "jpg")
}
\arguments{
\item{in_folder}{Path to an image folder}

\item{out_folder}{Path to an  folder to be made that will have the images out}

\item{width}{max width in pixels of the output crops}

\item{height}{max height in pixels of the output crops}

\item{buff}{buffer to overlap crop by x pixels}

\item{file_type}{file extentions}
}
\value{
not much really.  just a list of names of images in
}
\description{
Tile images in a directory
}
\author{
Abram Fleishman \email{abram@conservationmetics.com}
}
