% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vott_json2df.R
\name{vott_json2df}
\alias{vott_json2df}
\title{Read VoTT JSON as data.frame}
\usage{
vott_json2df(json_path)
}
\arguments{
\item{json_path}{Path to a VoTT json file}
}
\value{
a data.frame in long format with one row per bounding box
}
\description{
Tracks a VoTT JSON file and reads the tags and images to build a tidy data.frame
}
\author{
Abram Fleishman \email{abram@conservationmetics.com}
}
